requirements.txt
alembic==1.10.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-Sessions==0.1.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
